Working version of SSCA2 benchmark for kernel 1 -singleTM flag problems for more than 1 thread still exists, some problem in TRANSREAD, Amem_cpy()

+++ Robust/src/Benchmarks/SingleTM/SSCA2/VList.java
@@ -1,10 +1,9 @@
 /* Vertex list returned by Kernel 3 */
 public class VList {
-  public long num;
-  public long depth;
+  public int num;
+  public int depth;
 
   public VList() {
 
   }
-
 }
